--- a/src/laser_polio/model.py
+++ b/src/laser_polio/model.py
@@ -192,12 +192,12 @@ class SEIR_ABM:
                     self.t += 1
 
                     # Early stopping rule
-                    # if self.should_stop:
-                    #     if self.verbose >= 1:
-                    #         sc.printyellow(
-                    #             f"[SEIR_ABM] Early stopping at t={self.t}: no E/I and no future seed_schedule events. This stops all components (e.g., no births, deaths, or vaccination)"
-                    #         )
-                    #     break
+                    if self.should_stop:
+                        if self.verbose >= 1:
+                            sc.printyellow(
+                                f"[SEIR_ABM] Early stopping at t={self.t}: no E/I and no future seed_schedule events. This stops all components (e.g., no births, deaths, or vaccination)"
+                            )
+                        break
 
                 bar()  # Update the progress bar
         if self.verbose >= 1:
@@ -1270,12 +1270,7 @@ class VitalDynamics_ABM:
         if t % self.step_size != 0:
             return
 
-        # # 1) Vectorized mask of all alive people
-        # alive = self.people.disease_state >= 0
-
-        # # 2) Count how many alive in each node in one pass
-        # node_ids_alive = self.people.node_id[alive]
-        # alive_count_by_node = np.bincount(node_ids_alive, minlength=len(self.nodes))
+        # 1) Get vital statistics - alive and newly deceased
         num_nodes = len(self.nodes)
         tl_alive = np.zeros((nb.get_num_threads(), num_nodes), dtype=np.int32)
         tl_dying = np.zeros((nb.get_num_threads(), num_nodes), dtype=np.int32)
@@ -1294,47 +1289,34 @@ class VitalDynamics_ABM:
             deaths_count_by_node,
         )
 
-        # 3) Compute births node by node, but without big boolean masks
-        for node in self.nodes:
-            expected_births = self.step_size * self.birth_rate * alive_count_by_node[node]
+        # 2) Compute births
+        expected_births = self.step_size * self.birth_rate * alive_count_by_node
+        birth_integer = expected_births.astype(np.int32)
+        birth_fraction = expected_births - birth_integer
+        birth_rand = np.random.binomial(1, birth_fraction)  # Bernoulli draw
+        births = birth_integer + birth_rand
 
-            # Integer part plus probabilistic fractional part
-            birth_integer = int(expected_births)
-            birth_fraction = expected_births - birth_integer
-            birth_rand = np.random.binomial(1, birth_fraction)  # Bernoulli draw
-            births = birth_integer + birth_rand
+        if (total_births := births.sum()) > 0:
+            start, end = self.people.add(total_births)
 
-            # If births occur, add them to the population
-            if births > 0:
-                start, end = self.people.add(births)
+            dobs = self.people.date_of_birth[start:end]
+            dods = self.people.date_of_death[start:end]
 
-                newborn_ages = np.zeros(births, dtype=np.int32)
-                lifespans = self.death_estimator.predict_age_at_death(newborn_ages, max_year=100)
+            dobs[:] = 0  # temporarily
+            dods[:] = self.death_estimator.predict_age_at_death(dobs, max_year=100)
+            dobs[:] = t  # now set to current time
+            self.people.disease_state[start:end] = 0
+            dods[:] += t  # offset by current time
+            # assign node IDs to newborns
+            self.people.node_id[start:end] = np.repeat(np.arange(num_nodes), births)
+            if any(isinstance(component, RI_ABM) for component in self.sim.components):
+                self.people.ri_timer[start:end] = 182
 
-                self.people.date_of_birth[start:end] = t
-                self.people.disease_state[start:end] = 0
-                self.people.date_of_death[start:end] = lifespans + t
-                self.people.node_id[start:end] = node
-                if any(isinstance(component, RI_ABM) for component in self.sim.components):
-                    self.people.ri_timer[start:end] = 182
+            self.results.births[t] = births
 
-                self.results.births[t, node] = births
-
-                # # 4) Now handle deaths, again in a vectorized way
-                # #    People die if they're alive and their date_of_death <= t
-                # dying = alive & (self.people.date_of_death <= t)
-
-                # # Count how many are dying in each node
-                # node_ids_dying = self.people.node_id[dying]
-                # deaths_count_by_node = np.bincount(node_ids_dying, minlength=len(self.nodes))
-
-                # # Mark them dead
-                # self.people.disease_state[dying] = -1
-
-                # # 5) Store the death counts
-                # for node in self.nodes:
-                #     self.results.deaths[t, node] = deaths_count_by_node[node]
-                self.results.deaths[t] = deaths_count_by_node
+        # 3) Store the death counts
+        # Actual "death" handled in get_vital_statistics() as we count newly deceased
+        self.results.deaths[t] = deaths_count_by_node
 
         return