Added Numba-fied `get_vital_statistics()` function to accumulate alive and dead statistics in parallel.
Convert node-by-node births handling to vectorized NumPy (this is only at #nodes scale, so not bothering with Numba).

Speed up on MacBook (`demo_nigeria.py` 100% scaling, 2.5% initial prevalence, 3 years):

483 seconds (8:03) &rarr; 251 seconds (4:11)

Speed up on Windows: (`demo_nigeria.py` 100% scaling, 1% initial prevalence, 1 year):

582 seconds (9:42) &rarr; 371 seconds (6:01)